feat: 가용재고 Double -> Int, swagger-ui summary added

+++ src/main/kotlin/com/smartfoodnet/fnproduct/order/dto/ConfirmOrderModel.kt
@@ -45,6 +45,6 @@ data class ConfirmOrderModel @QueryProjection constructor(
     val collectedAt: LocalDateTime?
 ) {
     val mappedFlag: Boolean = basicProductId != null
-    var availableQuantity: Double = -1.0
+    var availableQuantity: Int = -1
     val mappedQuantityCalc: Int = (mappedQuantity ?: 0) * (quantity ?: 0)
 }
